miscellaneous state fixes

+++ internal/db/postgres/schema.sql
@@ -178,7 +178,7 @@ CREATE TABLE messages (
     model TEXT,
     agent TEXT,
     token_count BIGINT,
-    cost REAL,
+    cost DOUBLE PRECISION,
     workflow_id TEXT,
     run_id TEXT,
     node_id TEXT,
